refactor(ui): extract shared table logic into DeviceList and EmulsionTable components
- Add DeviceList component; all four device pages delegate to it
- Add EmulsionTable component with optional #actions slot; all emulsion pages delegate to it
- Migrate film/index.vue to use existing FilmInventoryTable
